+++ trixnity-messenger-compose-view/src/commonMain/kotlin/de/connect2x/messenger/compose/view/profiles/ProfileCreation.kt
@@ -1,5 +1,6 @@
 package de.connect2x.messenger.compose.view.profiles
 
+import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.size
@@ -11,18 +12,17 @@ import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.text.capitalize
-import androidx.compose.ui.text.intl.Locale
 import androidx.compose.ui.unit.dp
 import de.connect2x.messenger.compose.view.DI
 import de.connect2x.messenger.compose.view.collectAsTextFieldValueState
-import de.connect2x.messenger.compose.view.common.MessengerModal
-import de.connect2x.messenger.compose.view.common.MessengerModalButtonRow
-import de.connect2x.messenger.compose.view.common.MessengerModalContent
 import de.connect2x.messenger.compose.view.get
 import de.connect2x.messenger.compose.view.i18n.I18nView
 import de.connect2x.messenger.compose.view.theme.components
+import de.connect2x.messenger.compose.view.theme.components.ModalDialogContent
+import de.connect2x.messenger.compose.view.theme.components.ModalDialogFooter
+import de.connect2x.messenger.compose.view.theme.components.ModalDialogHeader
 import de.connect2x.messenger.compose.view.theme.components.ThemedButton
+import de.connect2x.messenger.compose.view.theme.components.ThemedModalDialog
 import de.connect2x.trixnity.messenger.multi.ProfileCreationViewModel
 
 interface ProfileCreationView {
@@ -43,47 +43,45 @@ class ProfileCreationViewImpl : ProfileCreationView {
         val error = profileCreationViewModel.error.collectAsState().value
         val canCreateProfile = profileCreationViewModel.canCreateProfile.collectAsState().value
 
-        MessengerModal(
-            onDismiss = onFinish,
-            title = i18n.createProfileHeader()
-        ) {
-            MessengerModalContent {
-                Text(i18n.createProfileSelectName(), style = MaterialTheme.typography.titleMedium)
-                Spacer(Modifier.size(10.dp))
-                OutlinedTextField(
-                    profileName,
-                    { profileName = it },
-                    modifier = Modifier.fillMaxWidth(),
-                    isError = error != null,
-                )
-                Spacer(Modifier.size(5.dp))
-                error?.let {
+        ThemedModalDialog(onFinish) {
+            ModalDialogHeader {
+                Text(i18n.createProfileHeader())
+            }
+            ModalDialogContent {
+                Column {
+                    Text(i18n.createProfileSelectName(), style = MaterialTheme.typography.titleMedium)
+                    Spacer(Modifier.size(10.dp))
+                    OutlinedTextField(
+                        profileName,
+                        { profileName = it },
+                        modifier = Modifier.fillMaxWidth(),
+                        isError = error != null,
+                    )
                     Spacer(Modifier.size(5.dp))
-                    Text(color = MaterialTheme.colorScheme.error, text = it)
+                    error?.let {
+                        Spacer(Modifier.size(5.dp))
+                        Text(color = MaterialTheme.colorScheme.error, text = it)
+                    }
                 }
             }
-            MessengerModalButtonRow(
-                button1 = {
-                    ThemedButton(
-                        style = MaterialTheme.components.commonButton,
-                        onClick = { onFinish() },
-                    ) {
-                        Text(i18n.commonCancel().capitalize(Locale.current))
-                    }
-                },
-                button2 = {
-                    ThemedButton(
-                        style = MaterialTheme.components.primaryButton,
-                        onClick = {
-                            profileCreationViewModel.createProfile()
-                            onFinish()
-                        },
-                        enabled = canCreateProfile,
-                    ) {
-                        Text(i18n.createProfileAction())
-                    }
+            ModalDialogFooter {
+                ThemedButton(
+                    style = MaterialTheme.components.commonButton,
+                    onClick = { onFinish() },
+                ) {
+                    Text(i18n.actionCancel())
                 }
-            )
+                ThemedButton(
+                    style = MaterialTheme.components.primaryButton,
+                    onClick = {
+                        profileCreationViewModel.createProfile()
+                        onFinish()
+                    },
+                    enabled = canCreateProfile,
+                ) {
+                    Text(i18n.createProfileAction())
+                }
+            }
         }
     }
 }