+++ trixnity-messenger-compose-view/src/commonMain/kotlin/de/connect2x/messenger/compose/view/theme/components/ThemedSurface.kt
@@ -4,6 +4,7 @@ import androidx.compose.foundation.BorderStroke
 import androidx.compose.foundation.background
 import androidx.compose.foundation.border
 import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.BoxScope
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.MaterialTheme
@@ -59,7 +60,7 @@ data class SurfaceStyle(
 fun ThemedSurface(
     modifier: Modifier = Modifier,
     style: SurfaceStyle,
-    content: @Composable () -> Unit
+    content: @Composable BoxScope.() -> Unit
 ) = Surface(
     modifier = modifier.padding(style.padding),
     shape = style.shape,